legacy tools used for downloading met forecast data off of the NOMADS server. Depends upon get_grib.pl and get_inv.pl, as well as on the grib2 tools, which were not included because they are not my code.
The main code is get_sref.R

get_sref.sh is a shell that calls get_sref.R and can be set up as a cron job

extract2driver is the code used to build met files for the ED2 model and is included for reference. The generic version was built for NARR data, the ebifarm version was for extraction from NOMADS SREF data at the EBI Energy Farm location.